+++ src/lib/utils/info.txt
@@ -9,7 +9,6 @@ assert.h
 bswap.h
 calendar.h
 charset.h
-cpuid.h
 compiler.h
 data_src.h
 database.h
@@ -45,3 +44,7 @@ linux -> rt
 mingw -> ws2_32
 windows -> ws2_32.lib
 </libs>
+
+<requires>
+cpuid
+</requires>
